+++ connection.go
@@ -8,9 +8,9 @@ import (
 
 	fasthttpWebsocket "github.com/fasthttp/websocket"
 	"github.com/fraym/golog"
-	graphql "github.com/fraym/graphql-go"
 	"github.com/gofiber/contrib/websocket"
 	"github.com/gofiber/fiber/v2"
+	graphql "github.com/jesko-plitt/graphql-go"
 )
 
 var ErrWebsocketClosed = fmt.Errorf("websocket closed")