+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # v0.4.0
 
-- (bc) Use fraym/graphql-go instead of graphql-go/graphql
+- (bc) Use jesko-plitt/graphql-go instead of graphql-go/graphql
 
 # v0.3.3